+++ engine/apps/alerts/tasks/unsilence.py
@@ -18,9 +18,7 @@ def unsilence_task(alert_group_pk):
     task_logger.info(f"Start unsilence_task for alert_group {alert_group_pk}")
     with transaction.atomic():
         try:
-            alert_group = AlertGroup.unarchived_objects.filter(pk=alert_group_pk).select_for_update()[
-                0
-            ]  # Lock alert_group:
+            alert_group = AlertGroup.objects.filter(pk=alert_group_pk).select_for_update()[0]  # Lock alert_group:
         except IndexError:
             task_logger.info(f"unsilence_task. alert_group {alert_group_pk} doesn't exist")
             return