+++ engine/apps/alerts/tasks/wipe.py
@@ -10,6 +10,6 @@ from common.custom_celery_tasks import shared_dedicated_queue_retry_task
 def wipe(alert_group_pk, user_pk):
     AlertGroup = apps.get_model("alerts", "AlertGroup")
     User = apps.get_model("user_management", "User")
-    alert_group = AlertGroup.all_objects.filter(pk=alert_group_pk).first()
+    alert_group = AlertGroup.objects.filter(pk=alert_group_pk).first()
     user = User.objects.filter(pk=user_pk).first()
     alert_group.wipe_by_user(user)